refactor(backend): extract Base to db_base.py to break db ↔ models cycle
Sentrux's acyclicity score sat at 5000/10000 because of one static SCC:

  backend/app/models.py  ─ from .db import Base ─────────┐
                                                         │
  backend/app/db.py      ─ from app import models ───────┘
                          (lazy import inside init_db()
                           to register every ORM class
                           with Base.metadata before
                           Base.metadata.create_all)

Grimp / sentrux count function-local imports as edges, so the cycle was
real even though the runtime hop was deferred. Extracting `Base` into a
sibling module lets both files import from it without a back-edge:

  backend/app/db_base.py (new) ─ class Base(DeclarativeBase)
       ▲                ▲
       │                │
  app/db.py        app/models.py     (both downward, no cycle)

The lazy `from app import models` inside `init_db()` is gone too — every
api/* router imports `app.models.*` directly, so the metadata is fully
registered long before lifespan startup reaches `create_db_and_tables()`.
Alembic (alembic/env.py:21) and pytest (tests/conftest.py:18) keep their
own explicit side-effect imports because they don't load main.py.

Also excludes `electrobun/` from sentrux's snapshot — it's a desktop-shell
spike with its own tsconfig + vitest gates, and a separate 4-file cycle
inside the spike was previously counting against the main app's budget
(masked behind the same `max_cycles = 1` exemption this PR removes).

Score impact:
  quality signal      6026 → 6912   (+886)
  acyclicity (raw 1→0, score 5000 → 10000)
  bottleneck shifts to modularity (5896) — addressed in subsequent PRs
